# [internal] Generate `_go_internal_package` targets