--- a/src/python/pants/backend/go/goals/tailor.py
+++ b/src/python/pants/backend/go/goals/tailor.py
@@ -4,7 +4,7 @@
 import os
 from dataclasses import dataclass
 
-from pants.backend.go.target_types import GoModTarget, GoPackage
+from pants.backend.go.target_types import GoModTarget
 from pants.core.goals.tailor import (
     AllOwnedSources,
     PutativeTarget,
@@ -17,32 +17,6 @@ from pants.engine.internals.selectors import Get
 from pants.engine.rules import collect_rules, rule
 from pants.engine.unions import UnionRule
 from pants.util.logging import LogLevel
-
-
-@dataclass(frozen=True)
-class PutativeGoPackageTargetsRequest(PutativeTargetsRequest):
-    pass
-
-
-@rule(level=LogLevel.DEBUG, desc="Determine candidate Go `go_package` targets to create")
-async def find_putative_go_package_targets(
-    request: PutativeGoPackageTargetsRequest, all_owned_sources: AllOwnedSources
-) -> PutativeTargets:
-    all_go_files = await Get(Paths, PathGlobs, request.search_paths.path_globs("*.go"))
-    unowned_go_files = set(all_go_files.files) - set(all_owned_sources)
-
-    putative_targets = []
-    for dirname, filenames in group_by_dir(unowned_go_files).items():
-        putative_targets.append(
-            PutativeTarget.for_target_type(
-                GoPackage,
-                dirname,
-                os.path.basename(dirname),
-                sorted(filenames),
-            )
-        )
-
-    return PutativeTargets(putative_targets)
 
 
 @dataclass(frozen=True)
@@ -62,9 +36,9 @@ async def find_putative_go_mod_targets(
         putative_targets.append(
             PutativeTarget.for_target_type(
                 GoModTarget,
-                dirname,
-                os.path.basename(dirname),
-                sorted(filenames),
+                path=dirname,
+                name=os.path.basename(dirname),
+                triggering_sources=sorted(filenames),
             )
         )
 
@@ -72,8 +46,4 @@ async def find_putative_go_mod_targets(
 
 
 def rules():
-    return [
-        *collect_rules(),
-        UnionRule(PutativeTargetsRequest, PutativeGoPackageTargetsRequest),
-        UnionRule(PutativeTargetsRequest, PutativeGoModuleTargetsRequest),
-    ]
+    return [*collect_rules(), UnionRule(PutativeTargetsRequest, PutativeGoModuleTargetsRequest)]