--- a/src/python/pants/backend/go/lint/fmt.py
+++ b/src/python/pants/backend/go/lint/fmt.py
@@ -4,7 +4,7 @@
 from dataclasses import dataclass
 from typing import Iterable
 
-from pants.backend.go.target_types import GoSources
+from pants.backend.go.target_types import GoInternalPackageSourcesField
 from pants.core.goals.fmt import FmtResult, LanguageFmtResults, LanguageFmtTargets
 from pants.core.goals.style_request import StyleRequest
 from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
@@ -18,7 +18,7 @@ from pants.engine.unions import UnionMembership, UnionRule, union
 # While gofmt is one of those, this is more generic and can work with other formatters.
 @dataclass(frozen=True)
 class GoLangFmtTargets(LanguageFmtTargets):
-    required_fields = (GoSources,)
+    required_fields = (GoInternalPackageSourcesField,)
 
 
 @union
@@ -32,7 +32,9 @@ async def format_golang_targets(
 ) -> LanguageFmtResults:
     original_sources = await Get(
         SourceFiles,
-        SourceFilesRequest(target[GoSources] for target in go_fmt_targets.targets),
+        SourceFilesRequest(
+            target[GoInternalPackageSourcesField] for target in go_fmt_targets.targets
+        ),
     )
     prior_formatter_result = original_sources.snapshot