--- a/testprojects/src/go/pants_test/BUILD
+++ b/testprojects/src/go/pants_test/BUILD
@@ -1,5 +1,4 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-go_package()
 go_binary(name="bin")