--- a/testprojects/src/go/pants_test/bar/BUILD
+++ b/testprojects/src/go/pants_test/bar/BUILD
@@ -1,5 +1,0 @@
-# Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-go_package()
-