… src/python/pants/backend/go/goals/package_binary_integration_test.py
… src/python/pants/backend/go/goals/tailor_test.py
… src/python/pants/backend/go/goals/test_test.py
… src/python/pants/backend/go/lint/gofmt/rules_integration_test.py
… src/python/pants/backend/go/target_type_rules_test.py
… src/python/pants/backend/go/util_rules/build_go_pkg.py
… src/python/pants/backend/go/util_rules/build_go_pkg_test.py
… src/python/pants/backend/go/util_rules/go_mod_test.py
… src/python/pants/backend/go/util_rules/go_pkg.py
… src/python/pants/backend/go/util_rules/go_pkg_test.py